[maint] Rename files from trial_response to response
