+++ lib/src/response/response_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mdigits/src/response/response_widget.dart';
 import 'package:mdigits/src/common/default_appbar.dart';
-import 'package:mdigits/src/response/trial_response_controller.dart';
+import 'package:mdigits/src/response/response_controller.dart';
 
 /// Screen for collecting response from participant
 class TrialResponseView extends StatelessWidget {
